Live stream of a Feedly feed w/ Phoenix channel
This has been tested with the Hacker News feed, but should work with all
feeds. Clicking "Articles" will give you an HTML view of a list of 20 feed
items instead of the JSON view (commented out). The JavaScript will push a
message to the Phoenix channel every 5 minutes asking for new articles
to prepend to the list. No fancy animations or anything.

Note: The way the LivestreamChannel module gathers the necessary info for
the API request seems hacky. Instead of pulling data from the session,
I have embedded the data into the HTML/JS template. I could not find a
way to access the `Plug.Conn` object from the channel module, as in the
template and controller. Suggestions are welcome.